Better handling of the multi-monitor setup:
The user 'SimonHyde-BBC' reports:

When multiple monitors are attached X translates the incoming clicks
to cover the whole display area across all monitors. This causes real
problems if the overall display isn't rectangular, such as when 2
monitors are different resolutions. In this case X11 won't generate a click
off the monitors, instead moving it to the nearest pixel (in the X direction?)
which is on a monitor. This means that if you have a 1024x768 monitor to the
left of a 1920x1080 monitor, all clicks in the bottom-left corner will
actually come in as clicks with an X co-ordinate of 1024 (the start of the
taller 1920x1080 monitor).
To prevent this problem, we must first translate and scale the whole
co-ordinate space of the overall display width/height, into the co-ordinate
space of the monitor we're drawing our window on, that way all clicks will
be scaled to values X11 will actually return to our program.

I rearranged his code; credit to him for the original idea and code.